add: bottom navi - provider as jotai global
