--- apps/web/src/components/provider/BottomNavi/BottomNaviProvider/BottomNaviContext.tsx
@@ -1,11 +1,0 @@
-import { createContext } from 'react'
-
-export type BottomNaviType = {
-  setBottomNavi: (_visible: boolean) => void
-}
-
-const context = createContext<BottomNaviType>({
-  setBottomNavi: () => { },
-})
-
-export default context
