--- apps/web/src/components/provider/BottomNavi/BottomNaviProvider/index.ts
@@ -1,2 +1,0 @@
-export { BottomNaviProvider as default } from './BottomNaviProvider'
-export { default as BottomNaviContext } from './BottomNaviContext'
