--- apps/web/src/components/provider/BottomNavi/hooks/useBottomNavi.ts
@@ -1,13 +1,0 @@
-import { useEffect } from 'react'
-import useBottomNaviContext from './useBottomNaviContext'
-
-export default function useBottomNavi() {
-  const context = useBottomNaviContext()
-  useEffect(() => {
-    context.setBottomNavi(true)
-    return () => {
-      context.setBottomNavi(false)
-    }
-  }, [context])
-
-}
