--- apps/web/src/components/provider/BottomNavi/hooks/useBottomNaviContext.ts
@@ -1,10 +1,0 @@
-import { useContext } from 'react'
-import { BottomNaviContext } from '../BottomNaviProvider'
-
-export default function useBottomNaviContext() {
-  const context = useContext(BottomNaviContext)
-  if (!context) {
-    throw new Error('Wrong position')
-  }
-  return context
-}
