+++ apps/web/src/components/ui/BottomNavi/atom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai'
+
+export const bottomNaviAtom = atom<boolean>(false)
