+++ apps/web/src/components/ui/BottomNavi/component/BottomNaviArea.tsx
