+++ apps/web/src/components/ui/BottomNavi/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component/BottomNavi'
+export { default as useBottomNavi } from './hooks/useBottomNavi'
