+++ apps/web/src/pages/:selectDate/workout/:id.tsx
@@ -8,7 +8,7 @@ import { useGetScheduleById, useUpdateSchedule } from '@hooks/apollo/Schedule'
 import usePageTracker from '@hooks/usePageTracker'
 import { useScheduleHeaderMenu } from '@hooks/useScheduleMenu'
 import useAlert from '@provider/Alert/hooks/useAlert'
-import BottomNaviArea from '@provider/BottomNavi/component/BottomNaviArea'
+import BottomNaviArea from '@ui/BottomNavi/component/BottomNaviArea'
 import ExerciseDataList from '@ui/ExerciseData/ExerciseDataList'
 import useHeaderHandler from '@ui/Header/hooks/useHeaderHandler'
 import PresetNameInputDialog from '@ui/Preset/PresetNameInputDialog'
