+++ apps/web/src/pages/Main.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom'
 import { useGetScheduleByDate } from '@hooks/apollo/Schedule'
 import usePageTracker from '@hooks/usePageTracker'
 import { useScheduleActions } from '@hooks/useScheduleMenu'
-import useBottomNavi from '@provider/BottomNavi/hooks/useBottomNavi'
+import useBottomNavi from '@ui/BottomNavi/hooks/useBottomNavi'
 import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion'
 import useHeaderHandler from '@ui/Header/hooks/useHeaderHandler'
 import ScheduleDisplay from '@ui/Schedule/ScheduleDisplay'
